Fix middleMonthEquation returning undefined (closes #11)
The formatting of the `middleMonthEquation` was a bit off, and the
`return equation` line was actually inside the `if (this.get('middleMonthCount') > 0) {`
line. Meaning, if there was no middle month count, nothing was returned.

In JavaScript, if nothing is explicitly returned from a function, the
function returns `undefined`. Hence the error.